bella

+++ app/Http/Controllers/User/GroupController.php
@@ -38,36 +38,9 @@ class GroupController extends Controller {
         // potrebbe non essere la scelta migliore: vedremo in futuro cosa è meglio fare!
         // cancello solo tutte le relazioni user-group
         $group = Group::find($id);
-        $users = $group->users()->detach();
+        $users = $group->users()->wherePivot('user_id','=',Auth::id())->detach();
 
-        Group::destroy($id);
-        // un pò brutto da vedere ma è l'unica soluzione che mi viene in mente ora
-        //$groups = Group::all();
-        $groups = Group::paginate(50);
-
-        foreach ($groups as $group) {
-            $group->count = $group->users()->count();
-        }
-
-        return view('user.groups.index')->with(['message' => 'delete',
-                    'groups' => $groups]);
-    }
-
-    public function merda(Request $request) {
-        if (strlen($request['search_key']) > 0) {
-            $byName = Group::where('name', 'LIKE', "%{$request['search_key']}%")->get();
-            $byDescription = Group::where('description', 'LIKE', "%{$request['search_key']}%")->get();
-            $union = $byName->union($byDescription);
-
-            foreach ($union as $group) {
-                $group->count = $group->users()->count();
-            }
-
-
-            return view('user.groups.index')->with(['groups' => $union,
-                        'search' => true]);
-        }
-        return view('user.groups.index')->with('groups', Group::paginate(10));
+        return redirect(route('user.groups.index'));
     }
 
 }
